Add data scope as a subclass
